content/canvas/test/test_2d.imageData.object.readonly.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.object.readonly</title>
<!-- Testing: ImageData objects properties are read-only -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var imgdata = ctx.getImageData(0, 0, 10, 10);
var w = imgdata.width;
var h = imgdata.height;
var d = imgdata.data;
imgdata.width = 123;
imgdata.height = 123;
imgdata.data = [100,100,100,100];
todo(imgdata.width === w, "imgdata.width === w");
todo(imgdata.height === h, "imgdata.height === h");
todo(imgdata.data === d, "imgdata.data === d");
todo(imgdata.data[0] === 0, "imgdata.data[\""+(0)+"\"] === 0");
todo(imgdata.data[1] === 0, "imgdata.data[\""+(1)+"\"] === 0");
todo(imgdata.data[2] === 0, "imgdata.data[\""+(2)+"\"] === 0");
todo(imgdata.data[3] === 0, "imgdata.data[\""+(3)+"\"] === 0");

SimpleTest.finish();

});
</script>
